requirements.txt
astroid==2.5.6
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-SocketIO==4.3.1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.8.3
pyparsing==2.4.7
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
sqlite-orm==0.0.2b0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
